refactor: rename and reorganize process input interfaces for improved clarity and structure
