+++ components/interfaces/check_result.go
