+++ components/interfaces/managed_process.go
