+++ components/interfaces/process_input_config.go
@@ -1,0 +1,9 @@
+package interfaces
+
+type IProcessInputConfig interface {
+	GetWaitFor() bool
+	GetRestart() bool
+	GetProcessType() string
+	GetMetadata(key string) (any, bool)
+	SetMetadata(key string, value any)
+}
