+++ components/interfaces/process_input_runtime.go
@@ -1,0 +1,9 @@
+package interfaces
+
+import "reflect"
+
+type IProcessInputRuntimeBase[T any, P IProcessInput[T]] interface {
+	GetObjectType() reflect.Type
+	GetObject() *P
+	GetFunction() IValidationFunc[T]
+}
