+++ components/interfaces/process_input_system.go
@@ -1,0 +1,13 @@
+package interfaces
+
+import "os/exec"
+
+type IProcessInputSystemBase[T any] interface {
+	GetCommand() string
+	GetArgs() []string
+	GetPath() string
+	GetCmd() *exec.Cmd
+	BuildCmd() *exec.Cmd
+	Cmd() *exec.Cmd
+	GetProperties() map[string]interface{}
+}
