+++ components/interfaces/validation_result.go
@@ -1,0 +1,7 @@
+package interfaces
+
+type IValidationResult interface {
+	GetIsValid() bool
+	GetMessage() string
+	GetError() error
+}
